api-gateway/requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cffi==1.17.1
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
multidict==6.3.2
passlib==1.7.4
propcache==0.3.1
pyasn1==0.4.8
pycares==4.6.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.19.0
